[Periphery] update NonfungiblePositionManager

+++ src/periphery/contracts/base/LiquidityManagement.sol
@@ -52,7 +52,7 @@ abstract contract LiquidityManagement is IAlgebraMintCallback, PeripheryImmutabl
         AddLiquidityParams memory params
     )
         internal
-        returns (uint128 liquidity, uint256 actualLiquidity, uint256 amount0, uint256 amount1, IAlgebraPool pool)
+        returns (uint128 liquidity, uint128 actualLiquidity, uint256 amount0, uint256 amount1, IAlgebraPool pool)
     {
         PoolAddress.PoolKey memory poolKey = PoolAddress.PoolKey({token0: params.token0, token1: params.token1});
